Move suse manager settings to the settings context
