+++ lib/trento/settings/suse_manager_settings.ex
@@ -1,6 +1,6 @@
 defmodule Trento.Settings.SuseManagerSettings do
   @moduledoc """
-  Schema for software updates settings.
+  Schema for suse manager settings.
   """
 
   use Ecto.Schema
