+++ lib/trento/software_updates.ex
@@ -4,69 +4,8 @@ defmodule Trento.SoftwareUpdates do
   """
   require Logger
 
-  alias Trento.Support.DateService
-
-  alias Trento.Repo
   alias Trento.SoftwareUpdates.Discovery
-  alias Trento.Settings.SuseManagerSettings
-
-  @type software_update_settings_save_submission :: %{
-          url: String.t(),
-          username: String.t(),
-          password: String.t(),
-          ca_cert: String.t() | nil
-        }
-
-  @type software_update_settings_change_submission :: %{
-          url: String.t() | nil,
-          username: String.t() | nil,
-          password: String.t() | nil,
-          ca_cert: String.t() | nil
-        }
-
-  @spec get_settings :: {:ok, SuseManagerSettings.t()} | {:error, :settings_not_configured}
-  def get_settings do
-    settings = Repo.one(SuseManagerSettings.base_query())
-
-    if settings do
-      {:ok, settings}
-    else
-      {:error, :settings_not_configured}
-    end
-  end
-
-  @spec save_settings(software_update_settings_save_submission, module()) ::
-          {:ok, SuseManagerSettings.t()}
-          | {:error, :settings_already_configured}
-          | {:error, any()}
-  def save_settings(settings_submission, date_service \\ DateService) do
-    with {:ok, :settings_not_configured, settings} <- ensure_no_settings_configured() do
-      settings
-      |> save_or_update_settings(settings_submission, date_service)
-      |> log_error("Error while saving software updates settings")
-    end
-  end
-
-  @spec change_settings(software_update_settings_change_submission, module()) ::
-          {:ok, SuseManagerSettings.t()}
-          | {:error, :settings_not_configured}
-          | {:error, any()}
-  def change_settings(settings_submission, date_service \\ DateService) do
-    with {:ok, settings} <- get_settings() do
-      settings
-      |> save_or_update_settings(settings_submission, date_service)
-      |> log_error("Error while updating software updates settings")
-    end
-  end
-
-  @spec clear_settings :: :ok
-  def clear_settings do
-    Repo.delete_all(SuseManagerSettings.base_query())
-
-    Discovery.clear_software_updates_discoveries()
-
-    :ok
-  end
+  alias Trento.Settings
 
   @spec test_connection_settings :: :ok | {:error, :connection_test_failed}
   def test_connection_settings do
@@ -85,7 +24,7 @@ defmodule Trento.SoftwareUpdates do
 
   @spec run_discovery :: :ok | {:error, :settings_not_configured}
   def run_discovery do
-    case get_settings() do
+    case Settings.get_suse_manager_settings() do
       {:ok, _} ->
         Discovery.discover_software_updates()
         :ok
@@ -107,7 +46,7 @@ defmodule Trento.SoftwareUpdates do
              | :error_getting_packages
              | :max_login_retries_reached}
   def get_software_updates(host_id) do
-    with {:ok, _} <- get_settings(),
+    with {:ok, _} <- Settings.get_suse_manager_settings(),
          {:ok, relevant_patches, upgradable_packages} <- Discovery.get_discovery_result(host_id) do
       {:ok, %{relevant_patches: relevant_patches, upgradable_packages: upgradable_packages}}
     end
@@ -120,7 +59,7 @@ defmodule Trento.SoftwareUpdates do
              | :error_getting_patches
              | :max_login_retries_reached}
   def get_packages_patches(package_ids) do
-    with {:ok, _} <- get_settings() do
+    with {:ok, _} <- Settings.get_suse_manager_settings() do
       result =
         package_ids
         |> ParallelStream.map(fn package_id ->
@@ -134,51 +73,4 @@ defmodule Trento.SoftwareUpdates do
       {:ok, result}
     end
   end
-
-  defp ensure_no_settings_configured do
-    case Repo.one(SuseManagerSettings.base_query()) do
-      nil ->
-        {:ok, :settings_not_configured, nil}
-
-      %SuseManagerSettings{} ->
-        Logger.error("Error: software updates settings already configured")
-        {:error, :settings_already_configured}
-    end
-  end
-
-  defp save_or_update_settings(settings, settings_submission, date_service) do
-    result =
-      case settings do
-        nil ->
-          %SuseManagerSettings{}
-          |> SuseManagerSettings.changeset(settings_submission, date_service)
-          |> Repo.insert()
-
-        %SuseManagerSettings{} ->
-          settings
-          |> SuseManagerSettings.changeset(settings_submission, date_service)
-          |> Repo.update()
-      end
-
-    case result do
-      {:ok, _} = success ->
-        Discovery.clear()
-
-        Task.Supervisor.start_child(Trento.TasksSupervisor, fn ->
-          Discovery.discover_software_updates()
-        end)
-
-        success
-
-      {:error, _} = error ->
-        error
-    end
-  end
-
-  defp log_error({:error, _} = error, message) do
-    Logger.error("#{message}: #{inspect(error)}")
-    error
-  end
-
-  defp log_error(result, _), do: result
 end
